Use bash for bin/setup
That the ruby version is mostly shelling out is a hint that a shell language
might be a better choice. It's more direct, slightly faster, it sidesteps
chicken/egg problems when installing ruby, and writing things in bash is fun.

It's good to have an excuse to flex those shell muscles 💪